move helpful parse and compare method to utils class to be able to reuse it elsewhere
