+++ src/test/java/org/commcare/cases/test/CaseParseAndReadTest.java
@@ -1,5 +1,7 @@
 package org.commcare.cases.test;
 
+import static org.commcare.test.utilities.CaseTestUtils.compareCaseDbState;
+
 import org.commcare.cases.instance.CaseInstanceTreeElement;
 import org.commcare.test.utilities.CaseTestUtils;
 import org.commcare.test.utilities.TestInstanceInitializer;
@@ -72,43 +74,5 @@ public class CaseParseAndReadTest {
                                     String caseDbState) throws Exception {
         config.parseIntoSandbox(this.getClass().getResourceAsStream(inputTransactions), sandbox, false);
         compareCaseDbState(sandbox, getClass().getResourceAsStream(caseDbState));
-    }
-
-    private static void compareCaseDbState(MockUserDataSandbox sandbox, InputStream caseDbState)
-            throws IOException {
-        byte[] parsedDb = serializeCaseInstanceFromSandbox(sandbox);
-        Document parsed = XmlComparator.getDocumentFromStream(new ByteArrayInputStream(parsedDb));
-        Document loaded = XmlComparator.getDocumentFromStream(caseDbState);
-
-        try {
-            XmlComparator.isDOMEqual(parsed, loaded);
-        } catch(Exception e) {
-            System.out.print(new String(parsedDb));
-
-            //NOTE: The DOM's definitely don't match here, so the strings cannot be the same.
-            //The reason we are asserting equality is because the delta between the strings is
-            //likely to do a good job of contextualizing where the DOM's don't match.
-            Assert.assertEquals("CaseDB output did not match expected structure(" + e.getMessage() + ")", new String(dumpStream(caseDbState)), new String(parsedDb));
-        }
-    }
-
-    private static byte[] serializeCaseInstanceFromSandbox(MockUserDataSandbox sandbox) {
-        try {
-            ByteArrayOutputStream bos = new ByteArrayOutputStream();
-            DataModelSerializer s = new DataModelSerializer(bos, new TestInstanceInitializer(sandbox));
-
-            s.serialize(new ExternalDataInstance(CaseTestUtils.CASE_INSTANCE, CaseInstanceTreeElement.MODEL_NAME), null);
-
-            return bos.toByteArray();
-        } catch (IOException e) {
-            throw new RuntimeException(e.getMessage());
-        }
-    }
-
-
-    private static byte[] dumpStream(InputStream is) throws IOException {
-        ByteArrayOutputStream bos = new ByteArrayOutputStream();
-        StreamsUtil.writeFromInputToOutput(is, bos);
-        return bos.toByteArray();
     }
 }
